+++ client/src/app/api/auth/[...asgardeo]/route.ts
@@ -10,35 +10,23 @@ export async function GET(request: NextRequest) {
     // Generate state for security
     const state = crypto.randomUUID();
     
-    // Get the callback URL from query params
-    const callbackUrl = searchParams.get('callbackUrl');
-    
-    // Redirect to Asgardeo for authentication (keep redirect_uri consistent)
+    // Redirect to Asgardeo for authentication (redirect directly to /admin)
     const asgardeoLoginUrl = `${process.env.NEXT_PUBLIC_ASGARDEO_BASE_URL}/oauth2/authorize?` + 
       `response_type=code&` +
       `client_id=${process.env.NEXT_PUBLIC_ASGARDEO_CLIENT_ID}&` +
       `scope=${encodeURIComponent(process.env.NEXT_PUBLIC_ASGARDEO_SCOPES || 'openid profile')}&` +
-      `redirect_uri=${encodeURIComponent(`${process.env.NEXT_PUBLIC_APP_URL}/api/auth/callback`)}&` +
+      `redirect_uri=${encodeURIComponent(`${process.env.NEXT_PUBLIC_APP_URL}/admin`)}&` +
       `state=${state}`;
     
     const response = NextResponse.redirect(asgardeoLoginUrl);
     
-    // Store state and callback URL for verification
+    // Store state for verification
     response.cookies.set('oauth_state', state, {
       httpOnly: true,
       secure: process.env.NODE_ENV === 'production',
       sameSite: 'lax',
       maxAge: 60 * 10 // 10 minutes
     });
-    
-    if (callbackUrl) {
-      response.cookies.set('oauth_callback', callbackUrl, {
-        httpOnly: true,
-        secure: process.env.NODE_ENV === 'production',
-        sameSite: 'lax',
-        maxAge: 60 * 10 // 10 minutes
-      });
-    }
     
     return response;
   }
